+++ plugin/core/gvrender_core_svg.c
@@ -464,6 +464,15 @@ static void svg_textspan(GVJ_t * job, pointf p, textspan_t * span)
         gvputs(job, "\" y=\"");
         gvprintdouble(job, -p.y);
         gvputs(job, "\"");
+	if (span->angle != 0.0) {
+	    gvputs(job, " transform=\"rotate(");
+	    gvprintdouble(job, -span->angle);
+	    gvputs(job, " ");
+	    gvprintdouble(job, p.x);
+	    gvputs(job, " ");
+	    gvprintdouble(job, -p.y);
+	    gvputs(job, ")\"");
+	}
     }
     pA = span->font->postscript_alias;
     if (pA) {